services/gateway-service/requirements.txt
Django==5.2.5
uvicorn==0.35.0
python-dotenv==1.1.1
psycopg[binary]==3.2.9
djangorestframework==3.16.1
httpx==0.28.1
gunicorn==23.0.0

requests==2.32.5